refactor: implicit cast from double to float may be lossy

+++ src/main/java/emu/grasscutter/game/managers/stamina/StaminaManager.java
@@ -475,7 +475,7 @@ public class StaminaManager extends BasePlayerManager {
                 if (consumption.amount < 0 && isCharacterStamina) {
                     // Do not apply reduction factor when recovering stamina
                     if (player.getTeamManager().getTeamResonances().contains(10301)) {
-                        consumption.amount *= 0.85f;
+                        consumption.amount *= (int) 0.85f;
                     }
                 }
                 // Delay 1 seconds before starts recovering stamina
@@ -543,8 +543,8 @@ public class StaminaManager extends BasePlayerManager {
             consumption.amount = ConsumptionType.CLIMBING.amount;
         }
         // Climbing specific reductions
-        consumption.amount *= getFoodCostReductionFactor(ClimbFoodReductionMap);
-        consumption.amount *= getTalentCostReductionFactor(ClimbTalentReductionMap);
+        consumption.amount *= (int) getFoodCostReductionFactor(ClimbFoodReductionMap);
+        consumption.amount *= (int) getTalentCostReductionFactor(ClimbTalentReductionMap);
         return consumption;
     }
 
@@ -560,8 +560,8 @@ public class StaminaManager extends BasePlayerManager {
             consumption.amount = ConsumptionType.SWIM_DASH.amount;
         }
         // Swimming specific reductions
-        consumption.amount *= getFoodCostReductionFactor(SwimFoodReductionMap);
-        consumption.amount *= getTalentCostReductionFactor(SwimTalentReductionMap);
+        consumption.amount *= (int) getFoodCostReductionFactor(SwimFoodReductionMap);
+        consumption.amount *= (int) getTalentCostReductionFactor(SwimTalentReductionMap);
         return consumption;
     }
 
@@ -571,7 +571,7 @@ public class StaminaManager extends BasePlayerManager {
             consumption.type = ConsumptionType.DASH;
             consumption.amount = ConsumptionType.DASH.amount;
             // Dashing specific reductions
-            consumption.amount *= getFoodCostReductionFactor(DashFoodReductionMap);
+            consumption.amount *= (int) getFoodCostReductionFactor(DashFoodReductionMap);
         }
         return consumption;
     }
@@ -583,8 +583,8 @@ public class StaminaManager extends BasePlayerManager {
         }
         Consumption consumption = new Consumption(ConsumptionType.FLY);
         // Flying specific reductions
-        consumption.amount *= getFoodCostReductionFactor(FlyFoodReductionMap);
-        consumption.amount *= getTalentCostReductionFactor(FlyTalentReductionMap);
+        consumption.amount *= (int) getFoodCostReductionFactor(FlyFoodReductionMap);
+        consumption.amount *= (int) getTalentCostReductionFactor(FlyTalentReductionMap);
         return consumption;
     }
